tambah model hapus

+++ app/views/mahasiswa/index.php
@@ -17,8 +17,8 @@
                 <?php foreach( $data['mhs'] as $mhs){ ?>
                   <li class="list-group-item d-flex justify-content-between align-items-center">
                       <?= $mhs['nama']?>
-                      <a href="<?= BASEURL; ?>/mahasiswa/hapus/<?= $mhs['id']; ?>" class="badge bg-danger float-right ml-1">hapus</a>
                       <a href="<?= BASEURL; ?>/mahasiswa/detail/<?= $mhs['id']; ?>" class="badge bg-primary float-right ml-1">detail</a>
+                      <a href="<?= BASEURL; ?>/mahasiswa/hapus/<?= $mhs['id']; ?>" class="badge bg-danger float-right ml-1" onclick="return confirm('yakin?);">hapus</a>
                   </li>
                 
                 <?php } ?>
